refactor: Replace TudeeNavGraph with TudeeAppScaffold and streamline navigation structure
