+++ app/src/main/java/com/example/tudeeapp/MainActivity.kt
@@ -10,7 +10,7 @@ import androidx.activity.enableEdgeToEdge
 import androidx.annotation.RequiresApi
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import org.koin.androidx.viewmodel.ext.android.viewModel
-import com.example.tudeeapp.presentation.navigation.TudeeNavGraph
+import com.example.tudeeapp.presentation.navigation.TudeeAppScaffold
 
 class MainActivity : ComponentActivity() {
     @SuppressLint("SourceLockedOrientationActivity")
@@ -23,7 +23,7 @@ class MainActivity : ComponentActivity() {
         enableEdgeToEdge()
         requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_PORTRAIT
         setContent {
-            TudeeNavGraph()
+            TudeeAppScaffold()
         }
     }
