+++ app/src/main/java/com/example/tudeeapp/presentation/navigation/TudeeNavGraph.kt
@@ -2,38 +2,13 @@ package com.example.tudeeapp.presentation.navigation
 
 import android.os.Build
 import androidx.annotation.RequiresApi
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.core.tween
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
-import androidx.compose.foundation.layout.navigationBarsPadding
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.SideEffect
 import androidx.compose.runtime.compositionLocalOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.dialog
-import androidx.navigation.compose.rememberNavController
-import com.example.tudeeapp.data.source.local.sharedPreferences.AppPreferences
-import com.example.tudeeapp.presentation.common.components.SnackBar
-import com.example.tudeeapp.presentation.common.components.SnackBarState
-import com.example.tudeeapp.presentation.common.components.TudeeNavigationBar
-import com.example.tudeeapp.presentation.common.components.TudeeScaffold
-import com.example.tudeeapp.presentation.design_system.theme.Theme
-import com.example.tudeeapp.presentation.design_system.theme.TudeeTheme
 import com.example.tudeeapp.presentation.screen.categories.CategoriesScreen
 import com.example.tudeeapp.presentation.screen.categoriesForm.CategoryForm
 import com.example.tudeeapp.presentation.screen.categoryDetails.CategoryDetailsScreen
@@ -44,91 +19,31 @@ import com.example.tudeeapp.presentation.screen.splash.SplashScreen
 import com.example.tudeeapp.presentation.screen.task.TasksScreen
 import com.example.tudeeapp.presentation.screen.taskDetails.TaskDetailsScreen
 import com.example.tudeeapp.presentation.screen.taskManagement.TaskManagementBottomSheet
-import com.google.accompanist.systemuicontroller.rememberSystemUiController
-import kotlinx.coroutines.delay
 
 val LocalNavController = compositionLocalOf<NavHostController> { error("No Nav Controller Found") }
-val LocalSnackBarState = compositionLocalOf<SnackBarState> { error("No SnackBarState provided") }
-val LocalThemeState = compositionLocalOf<MutableState<TudeeThemeMode>> { error("No TaskManagementState provided") }
 
 @RequiresApi(Build.VERSION_CODES.O)
 @Composable
-fun TudeeNavGraph() {
-    val navController = rememberNavController()
-    val snackBarState = remember { SnackBarState() }
-    val context = LocalContext.current
-    val appPrefs = remember { AppPreferences(context) }
-    val themeMode = rememberSaveable {
-        mutableStateOf(if (appPrefs.isDarkTheme()) TudeeThemeMode.DARK else TudeeThemeMode.LIGHT)
-    }
-
-    val systemUiController = rememberSystemUiController()
-    val darkIcons = themeMode.value == TudeeThemeMode.LIGHT
-
-    SideEffect {
-        systemUiController.setSystemBarsColor(
-            color = Color.Transparent,
-            darkIcons = darkIcons
-        )
-    }
-
+fun TudeeNavGraph(navController: NavHostController) {
 
     CompositionLocalProvider(
         LocalNavController provides navController,
-        LocalSnackBarState provides snackBarState,
-        LocalThemeState provides themeMode
     ) {
-        TudeeTheme(
-            isDarkTheme = themeMode.value == TudeeThemeMode.DARK
+        NavHost(
+            navController = navController,
+            startDestination = Screens.Splash,
         ) {
-            TudeeScaffold(
-                modifier = Modifier.navigationBarsPadding(),
-                bottomBar = { TudeeNavigationBar(navController) },
-                contentBackground = Theme.colors.surfaceColors.surface
-            ) {
-                NavHost(
-                    navController = navController,
-                    startDestination = Screens.Splash,
-                ) {
-                    composable<Screens.Splash> { SplashScreen() }
-                    composable<Screens.OnBoarding> { OnBoardScreen(pages = onboardingPages()) }
-                    composable<Screens.Home> { HomeScreen() }
-                    composable<Screens.Tasks> { TasksScreen() }
-                    composable<Screens.Category> { CategoriesScreen() }
-                    dialog<Screens.TaskManagement> { TaskManagementBottomSheet() }
-                    dialog<Screens.TaskDetails> { TaskDetailsScreen() }
-                    composable<Screens.CategoryDetails> { CategoryDetailsScreen() }
-                    dialog<Screens.CategoryForm> {
-                        CategoryForm()
-                    }
-                }
-
-                if (snackBarState.isVisible) {
-                    AnimatedVisibility(
-                        visible = snackBarState.isVisible,
-                        enter = fadeIn(animationSpec = tween(snackBarState.durationMillis)),
-                        exit = fadeOut(animationSpec = tween(snackBarState.durationMillis))
-                    ) {
-                        SnackBar(
-                            Modifier
-                                .statusBarsPadding()
-                                .padding(start = 16.dp, end = 16.dp, top = 16.dp),
-                            text = snackBarState.message,
-                            isSuccess = snackBarState.isSuccess,
-                            onClick = { snackBarState.hide() }
-                        )
-                    }
-                    LaunchedEffect(Unit) {
-                        delay(snackBarState.durationMillis.toLong())
-                        snackBarState.hide()
-                    }
-                }
+            composable<Screens.Splash> { SplashScreen() }
+            composable<Screens.OnBoarding> { OnBoardScreen(pages = onboardingPages()) }
+            composable<Screens.Home> { HomeScreen() }
+            composable<Screens.Tasks> { TasksScreen() }
+            composable<Screens.Category> { CategoriesScreen() }
+            dialog<Screens.TaskManagement> { TaskManagementBottomSheet() }
+            dialog<Screens.TaskDetails> { TaskDetailsScreen() }
+            composable<Screens.CategoryDetails> { CategoryDetailsScreen() }
+            dialog<Screens.CategoryForm> {
+                CategoryForm()
             }
         }
     }
-}
-
-enum class TudeeThemeMode(val value: Boolean) {
-    DARK(true),
-    LIGHT(false)
 }
